+++ src/recv_udp.h
@@ -72,8 +72,29 @@ public:
      *                     operating system might not allow a buffer size
      *                     as big as the default.
      */
-    explicit udp_reader(
+    udp_reader(
         stream &owner,
+        const boost::asio::ip::udp::endpoint &endpoint,
+        std::size_t max_size = default_max_size,
+        std::size_t buffer_size = default_buffer_size);
+
+    /**
+     * Constructor using an existing socket. This allows socket options (e.g.,
+     * multicast subscriptions) to be set by the caller. The socket should not
+     * be bound.
+     *
+     * @param owner        Owning stream
+     * @param socket       Existing socket which will be taken over. It must
+     *                     use the same I/O service as @a owner.
+     * @param endpoint     Address on which to listen
+     * @param max_size     Maximum packet size that will be accepted.
+     * @param buffer_size  Requested socket buffer size. Note that the
+     *                     operating system might not allow a buffer size
+     *                     as big as the default.
+     */
+    udp_reader(
+        stream &owner,
+        boost::asio::ip::udp::socket &&socket,
         const boost::asio::ip::udp::endpoint &endpoint,
         std::size_t max_size = default_max_size,
         std::size_t buffer_size = default_buffer_size);